pyproject.toml
[tool.poetry]
name = "diss-db"
version = "0.1.0"
description = ""
authors = ["Savannah Scott <sscott2@eagles.bridgewater.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
asgiref = "3.5.2"
black = "22.10.0"
click = "8.1.3"
django = "4.1.3"
mypy-extensions = "0.4.3"
pathspec = "0.10.1"
platformdirs = "2.5.3"
psycopg2 = "2.9.5"
sqlparse = "0.4.3"
tomli = "2.0.1"
postgres = "^4.0"
django-tables2 = "^2.7.0"
django-filter = "23.5"
django-bootstrap3 = "^23.6"
djangorestframework = "^3.15.1"
httpie = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
